fix(build-vendor): make conditional

+++ packages/core/build-vendor.sh
@@ -2,8 +2,11 @@
 
 set -e
 
-cp -r ../../node_modules/@carbon/layout src/styles/vendor/@carbon
-cp -r ../../node_modules/@carbon/motion src/styles/vendor/@carbon
-cp -r ../../node_modules/@carbon/type src/styles/vendor/@carbon
-cp -r ../../node_modules/@carbon/themes src/styles/vendor/@carbon
-cp -r ../../node_modules/@carbon/colors src/styles/vendor/@carbon
+if [ -e ../../node_modules/@carbon/layout ]
+then
+	cp -r ../../node_modules/@carbon/layout src/styles/vendor/@carbon
+	cp -r ../../node_modules/@carbon/motion src/styles/vendor/@carbon
+	cp -r ../../node_modules/@carbon/type src/styles/vendor/@carbon
+	cp -r ../../node_modules/@carbon/themes src/styles/vendor/@carbon
+	cp -r ../../node_modules/@carbon/colors src/styles/vendor/@carbon
+fi
